rename project from org.scala-tools:scala-tools-server to alchim31.net:scalacs

+++ src/main/scala/net_alchim31_scalacs/compiler.scala
@@ -1,4 +1,4 @@
-package org.scala_tools.server
+package net_alchim31_scalacs
 
 import java.io.{BufferedOutputStream, File, FileOutputStream, PrintStream}
 import java.lang.{Runtime, System, Thread}
@@ -17,27 +17,65 @@ import java.util.regex.{Pattern, PatternSyntaxException}
 //TODO define CompilationGroup as Trait with Composite and Single implementation
 //TODO find a better name (eg, Builder, CompilerService)
 
+class CompileFeedback(
+  val runId : Long,
+  val out : CharSequence,
+  val isChanged : Boolean,
+  val isSuccess : Boolean
+) {
+  def +(o : CompileFeedback) = {
+    if (runId != o.runId) {
+      throw new IllegalArgumentException("can't add CompileFeedback from differend runId :" + runId + " != " + o.runId)
+    }
+    new CompileFeedback(
+      runId,
+      out.toString + o.out.toString,
+      isChanged || o.isChanged,
+      isSuccess && o.isSuccess
+    )
+  }
+}
+
 trait CompilerService {
   def reset() : CharSequence
   def clean() : CharSequence
-  def compile() : CharSequence
+  def compile(runId : Long, checkDeps : Boolean) : List[CompileFeedback]
 }
 
 class CompilerService4Group extends CompilerService {
   private
-  var _group : List[CompilerService] = Nil
+  var _group : List[CompilerService4Single] = Nil
 
-  def add(v : CompilerService) : this.type = {
+  def add(v : CompilerService4Single) : this.type = {
     // should generate new name if name already exist or replace existing ??
-    _group = v :: _group
+    _group = v :: _group.filter(_.cfg.name == v.cfg.name)
     this
   }
+
+  def remove(v : CompilerService4Single) : this.type = {
+    // should generate new name if name already exist or replace existing ??
+    _group = _group.filter(_.cfg.name == v.cfg.name)
+    this
+  }
+
   def size = _group.size
 
   // dispatching
   def reset() : CharSequence = _group.foldLeft("")((r,i) => r + i.reset().toString)
   def clean() : CharSequence = _group.foldLeft("")((r,i) => r + i.clean().toString)
-  def compile() : CharSequence = _group.foldLeft("")((r,i) => r + i.compile().toString)
+  def compile(runId : Long, checkDeps : Boolean) = _group.flatMap(_.compile(runId, checkDeps))//.removeDuplicates()
+
+  def findByName(p : Pattern) = _group.find(i => p.matcher(i.cfg.name).matches())
+  def findByTargetDir(f : File) = {
+    val apath = f.getCanonicalPath
+    _group.find(i => apath == i.cfg.targetDir.getCanonicalPath)
+  }
+
+  def findByExported(f : File) = {
+    val apath = f.getCanonicalPath
+    _group.find(i => i.cfg.exported.map( apath == _.getCanonicalPath).getOrElse(false))
+  }
+
 }
 
 class SingleConfig (
@@ -47,10 +85,11 @@ class SingleConfig (
   val sourceExcludes : List[Pattern],
   val targetDir : File,
   val classpath : List[File],
-  val additionalArgs : List[String]
+  val additionalArgs : List[String],
+  val exported : Option[File]
 )
 
-class CompilerService4Single(val cfg : SingleConfig) extends CompilerService {
+class CompilerService4Single(val cfg : SingleConfig, val allCompilerService : Option[CompilerService4Group]) extends CompilerService {
 
   val MaxCharge = 0.8
 
@@ -98,21 +137,64 @@ class CompilerService4Single(val cfg : SingleConfig) extends CompilerService {
   }
 
   def clean() : CharSequence = {
-    def delete(f : File) {
+    def delete(f : File) : Boolean = {
+      var back =  true
       if (f.isDirectory) {
         for (f <- f.listFiles()) {
-          delete(f)
+          back = delete(f) && back
         }
+      } else if (!f.getName.endsWith(".class")) {
+        back = false
       }
-      f.delete()
+      if (back) {
+        f.delete()
+      }
+      back
     }
     delete(cfg.targetDir)
-    "[Compile server has deleted " + cfg.targetDir + "]\n"
+    "[Compile server has deleted classes into " + cfg.targetDir + "]\n"
   }
 
+  var _lastCompileFeedback = new CompileFeedback(
+    java.lang.Long.MIN_VALUE,
+    "",
+    false,
+    true
+  )
+
+
   //TODO update reporter from compiler to use a new out for each action
-  def compile() :  CharSequence = {
+  //TODO use the GraphNode mixin
+  //TODO use LoggerMessage structure instead of CharSequence to store status, messages,...
+  def compile(runId : Long, checkDeps : Boolean) : List[CompileFeedback] = {
     import java.io.{PrintWriter, StringWriter, StringReader, BufferedReader}
+
+    if (runId <= _lastCompileFeedback.runId) {
+      throw new Exception("old runId ('" + runId +"') requested after '" + _lastCompileFeedback.runId +"'")
+    }
+
+    var back : List[CompileFeedback] = Nil
+
+    if (runId == _lastCompileFeedback.runId) {
+      return List(_lastCompileFeedback)
+    }
+
+    // request uptodate dependencies :
+    val classpath = allCompilerService match {
+      case None => cfg.classpath
+      case Some(acs) => cfg.classpath.map { f =>
+        acs.findByExported(f) match {
+          case Some(prj) =>  back = back ::: prj.compile(runId, checkDeps); prj.cfg.targetDir
+          case None => acs.findByTargetDir(f) match {
+            case Some(prj) =>  back = back ::: prj.compile(runId, checkDeps); prj.cfg.targetDir
+            case None => (f.lastModified() > _lastCompileFeedback.runId) match {
+              case true => back = new CompileFeedback(runId, "file " + f +" updated since last compile", true, true) :: back; f
+              case false => f
+            }
+          }
+        }
+      }
+    }
 
     if (!cfg.targetDir.exists()) {
       cfg.targetDir.mkdirs();
@@ -126,19 +208,27 @@ class CompilerService4Single(val cfg : SingleConfig) extends CompilerService {
     }
 
     val now = System.currentTimeMillis()
-    val files = findFilesToCompile(lastCompileAt)
-
-    if (files.isEmpty) {
-      return "Nothing to compile - all classes are up to date\n"
+    val classpathIsChanged = back.foldLeft(false)((r,i) => r || i.isChanged)
+    val files = classpathIsChanged match {
+      case false => findFilesToCompile(lastCompileAt)
+      case true => reset() ; findFilesToCompile(java.lang.Long.MIN_VALUE)
     }
 
-    val back = new StringWriter()
-    val out = new PrintWriter(back)
-    val in = new BufferedReader(new StringReader(""))
+    if (files.isEmpty) {
+      return new CompileFeedback(
+        runId,
+        "Nothing to compile - all classes are up to date",
+        false,
+        true
+      ) :: back
+    }
+//....
+    val outStr = new StringWriter()
+    val out = new PrintWriter(outStr)
 
     try {
       def error(msg: String) {
-        out.println(/*new Position*/ FakePos("fsc"), msg + "\n  fsc -help  gives more information")
+        out.println(/*new Position*/ FakePos("hsc"), msg + "\n  hsc -help  gives more information")
       }
 
       val args = cfg.additionalArgs :::
@@ -152,7 +242,7 @@ class CompilerService4Single(val cfg : SingleConfig) extends CompilerService {
       out.println("Compiling " + files.length + " source files to " + cfg.targetDir)
 
       val command = new OfflineCompilerCommand(args, new Settings(error), error, false)
-      val reporter = new ConsoleReporter(command.settings, in, out) {
+      val reporter = new CompilerReporter(command.settings, out) {
         // disable prompts, so that compile server cannot block
         override def displayPrompt = ()
       }
@@ -219,3 +309,45 @@ class CompilerService4Single(val cfg : SingleConfig) extends CompilerService {
     }
   }
 }
+
+abstract class Projects {
+
+  /**
+   * Update the definition of existing project with the same name
+   * Update projects dependency graph
+   */
+  def add(v : ProjectRawInfo)
+
+  /**
+   * visit the list of project sorted by dependency
+   */
+  def visit(f : ProjectRawInfo => Unit)
+}
+
+case class VFile(f : File, link : Option[File])
+class VFileList extends immutable.List[VFile] {
+  def realPath : List[File] = map{ vf =>
+    vf.link.filter(_.exists).getOrElse(vf.f)
+  }
+
+  def realPathStr = realPath.map(f.getCanonicalPath).mkString(File.separatorPath)
+
+  def update(vf : VFile) = {
+    (head.f == vf.f) match {
+      case true => vf :: tail
+      case false => head :: update(tail)
+    }
+  }
+}
+
+trait TypeInfo {
+  def src : File
+  def dotclass : File
+  def directDep : Set[TypeName]
+}
+
+class SrcDependencyTools {
+  def srcToClass(src : File)
+  def classToClassFile
+}
+
